pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robot"
version = "0.1.0"
description = "Unified Python interface for CLI-based coding agents"
readme = "README.md"
license = "CC0-1.0"
authors = [
    {name = "Philip Orange", email = "git@philiporange.com"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "prompt_toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
minimal = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
robot = "robot.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
